requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pypdf>=4.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
pandas>=2.0.0
